orders

+++ back/Models/Order.cs
@@ -8,7 +8,7 @@ namespace CarBon.Models
     {
         public int id { get; set; }
         public string title { get; set; }
-        public int price { get; set; }
+        public long price { get; set; }
         public string image { get; set; }
     }
 }
